+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ## Unreleased
 
 * refactor: allow to change max request body JSON nesting depth via injection
+* rollback: lock pagination in list requests to page-based again, as others are not yet supported
 
 ## 0.24.35 - 2024-01-09